requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
currencyapicom==0.1.1
Django==4.2
django-celery-beat==2.5.0
django-parler==2.3
django-timezone-field==5.0
everapi==0.1
idna==3.4
kombu==5.2.4
Pillow==9.5.0
prompt-toolkit==3.0.38
python-crontab==2.7.1
python-dateutil==2.8.2
python-gettext==5.0
pytz==2023.3
redis==4.5.4
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
